结构重构，数据库映射 树结构 加入大量工具类 初始化操作

+++ studentSys/src/main/java/com/hudongwx/studentSys/common/BaseStudent.java
@@ -9,11 +9,11 @@ import com.jfinal.plugin.activerecord.IBean;
 @SuppressWarnings("serial")
 public abstract class BaseStudent<M extends BaseStudent<M>> extends Model<M> implements IBean {
 
-	public void setId(java.lang.Integer id) {
+	public void setId(java.lang.Long id) {
 		set("id", id);
 	}
 
-	public java.lang.Integer getId() {
+	public java.lang.Long getId() {
 		return get("id");
 	}
 
@@ -41,11 +41,11 @@ public abstract class BaseStudent<M extends BaseStudent<M>> extends Model<M> imp
 		return get("tutorId");
 	}
 
-	public void setSubject(java.lang.Integer subject) {
+	public void setSubject(java.lang.Long subject) {
 		set("subject", subject);
 	}
 
-	public java.lang.Integer getSubject() {
+	public java.lang.Long getSubject() {
 		return get("subject");
 	}
